validateToken updated

+++ controller/contactsController.js
@@ -2,7 +2,7 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
 
   if (!contacts) {
     res.status(404);
@@ -24,6 +24,7 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
 
   res.status(201).json(contact);
@@ -48,6 +49,11 @@ const updateContact = asyncHandler(async (req, res) => {
     throw new Error("No Contacts Found");
   }
 
+  if (req.user.id !== contact.user_id.toString()) {
+    res.status(401);
+    throw new Error("User dont have permission to update");
+  }
+
   const updatedContact = Contact.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -61,6 +67,11 @@ const deleteContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error("No Contacts Found");
+  }
+
+  if (req.user.id !== contact.user_id.toString()) {
+    res.status(401);
+    throw new Error("User dont have permission to delete");
   }
 
   await Contact.deleteOne({ _id: req.params.id });
